Support optional arguments to ActiveStorage::Service#upload
Rails 5.2.1.1 was released to fix a couple of CVEs, one of which
affected ActiveStorage.
See https://weblog.rubyonrails.org/2018/11/27/Rails-4-2-5-0-5-1-5-2-have-been-released/

As part of the fix, optional service metadata can be supplied to the upload
method.

This commit collects up any extra keyword arguments, and ignores them. But at
least is API compatible with the ActiveStorage::Service